+++ src/network/networkHeader.js
@@ -49,6 +49,8 @@ export const NetworkHeader = GObject.registerClass({
         
         Config.connect(this, 'changed::visible', this.resetMaxWidths.bind(this));
         Config.connect(this, 'changed::network-header-io', this.resetMaxWidths.bind(this));
+        Config.connect(this, 'changed::headers-font-family', this.resetMaxWidths.bind(this));
+        Config.connect(this, 'changed::headers-font-size', this.resetMaxWidths.bind(this));
     }
     
     resetMaxWidths() {
@@ -139,7 +141,8 @@ export const NetworkHeader = GObject.registerClass({
     buildSpeed() {
         this.speedContainer = new St.BoxLayout({
             style_class: 'astra-monitor-header-speed-container',
-            x_align: Clutter.ActorAlign.CENTER,
+            x_align: Clutter.ActorAlign.START,
+            y_align: Clutter.ActorAlign.CENTER,
             y_expand: true,
             vertical: true,
             width: 1